refactor(discord-commands): create an internal function to search something inside a message
this refactor is not reducing the complexity it only move it to only one place which then could be rewrite more on another PR
fixes #778.
