+++ src/features/discord/messages/interfaces/commands/discord-extract-from-command-callback-data.ts
@@ -1,0 +1,7 @@
+import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
+
+export interface IDiscordExtractFromCommandCallbackData {
+  command: DiscordMessageCommandEnum;
+  message: string;
+  prefix: string;
+}
